+++ resources/views/admin/teacher_management/index.blade.php
@@ -16,7 +16,9 @@
 <div class="card">
   <div class="card-header d-flex justify-content-between align-items-center">
     <h3 class="mb-0 text-primary">Teacher List</h3>
-    <a href="{{ route('admin.teacher.create') }}" class="btn btn-primary btn-sm">+ Add Teacher</a>
+    @if (hasPermissionByChild('create_teacher'))
+      <a href="{{ route('admin.teacher.create') }}" class="btn btn-primary btn-sm">+ Add Teacher</a>
+    @endif
   </div>
 
   <div class="px-3 py-2">
@@ -36,13 +38,15 @@
                   <a href="{{ url()->current() }}" class="btn btn-sm btn-light" data-toggle="tooltip" title="Clear filter">
                     <i class="tf-icons ri-close-line"></i>
                   </a>
-                  <div class="d-md-flex justify-content-between align-items-center dt-layout-start">
-                    <a href="{{ route('admin.teacher.export', ['keyword' => request()->input('keyword')]) }}" 
-                      class="btn buttons-collection btn-outline-secondary waves-effect" 
-                      data-toggle="tooltip" title="Export Data">
-                      Export<i class="tf-icons ri-download-line"></i>
-                    </a>
-                  </div>
+                  @if (hasPermissionByChild('export_teacher_list'))
+                    <div class="d-md-flex justify-content-between align-items-center dt-layout-start">
+                      <a href="{{ route('admin.teacher.export', ['keyword' => request()->input('keyword')]) }}" 
+                        class="btn buttons-collection btn-outline-secondary waves-effect" 
+                        data-toggle="tooltip" title="Export Data">
+                        Export<i class="tf-icons ri-download-line"></i>
+                      </a>
+                    </div>
+                  @endif
                 </div>
               </div>
             </div>
@@ -74,7 +78,8 @@
               $subjects = [];
               foreach ($item->teacherSubjects as $teacherSubject) {
                 $clasName = $teacherSubject->classList->class;
-                $subjectName = $teacherSubject->subjectList->sub_name;
+                // $subjectName = $teacherSubject->subjectList->sub_name;
+                $subjectName = $teacherSubject->subjectList->sub_name ?? 'N/A';
 
                 if (!in_array($clasName, $classes)) {
                   $classes[] = $clasName;
@@ -96,21 +101,35 @@
                     <label class="form-check-label" for="customSwitch{{$item->id}}"></label>
                   </div>
               </td>
-              {{-- Edit and delete --}}
-              <td>
+              {{-- View, Edit and delete --}}
+              <td>               
                 <div class="btn-group" role="group" aria-label="Action Buttons">
-                  <a href="{{ route('admin.teacher.show', $item->id) }}"  class="btn btn-sm btn-icon btn-outline-success"         
-                    data-bs-toggle="tooltip" title="View">                  
-                    <i class="ri-eye-line"></i>
-                  </a>
-                  <a href="{{ route('admin.teacher.edit', $item->id) }}" class="btn btn-sm btn-icon btn-outline-dark"                     
-                    data-bs-toggle="tooltip"  title="Edit">
-                    <i class="ri-pencil-line"></i>
-                  </a>
-                  <a href="javascript:void(0);" class="btn btn-sm btn-icon btn-outline-danger" onclick="deleteTeacher({{ $item->id }})"     
-                    data-bs-toggle="tooltip" title="Delete">
-                    <i class="ri-delete-bin-6-line"></i>
-                  </a>
+                  @if (hasPermissionByChild('details_teacher'))
+                    <div>
+                      <a href="{{ route('admin.teacher.show', $item->id) }}"  class="btn btn-sm btn-icon btn-outline-success"         
+                        data-bs-toggle="tooltip" title="View">                  
+                        <i class="ri-eye-line"></i>
+                      </a>
+                    </div>
+                  @endif
+                 
+                  @if (hasPermissionByChild('edit_teacher'))
+                    <div>
+                      <a href="{{ route('admin.teacher.edit', $item->id) }}" class="btn btn-sm btn-icon btn-outline-dark"                     
+                        data-bs-toggle="tooltip"  title="Edit">
+                        <i class="ri-pencil-line"></i>
+                      </a>
+                    </div>
+                  @endif
+                  
+                  @if (hasPermissionByChild('delete_teacher'))
+                    <div>
+                      <a href="javascript:void(0);" class="btn btn-sm btn-icon btn-outline-danger" onclick="deleteTeacher({{ $item->id }})"     
+                        data-bs-toggle="tooltip" title="Delete">
+                        <i class="ri-delete-bin-6-line"></i>
+                      </a>
+                    </div> 
+                  @endif               
                 </div>
               </td> 
             </tr>
